change for delete proof

+++ src/main/java/starlight/backend/proof/service/impl/ProofServiceImpl.java
@@ -92,7 +92,10 @@ public class ProofServiceImpl implements ProofServiceInterface {
     }
 
     @Override
-    public void deleteProof(long talentId, long proofId) {
+    public void deleteProof(long talentId, long proofId, Authentication auth) {
+        if (!securityService.checkingLoggedAndTokenValid(talentId, auth)) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
         ProofEntity proof = em.find(ProofEntity.class, proofId);
         proof.setUser(null);
         em.remove(proof);
